Split Sensor Subsystem into three subsystems
